requirements.txt
# Core ML
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
xgboost==2.0.3
shap==0.45.0

# MLflow
mlflow==2.13.0
boto3==1.34.84          # for S3 artefact store

# Spark / Delta (local dev)
pyspark==3.5.1
delta-spark==3.1.0

# Serving
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-jose[cryptography]==3.3.0
slowapi==0.1.9

# Monitoring
scipy==1.13.0

# Testing
pytest==8.2.0
pytest-cov==5.0.0
pytest-asyncio==0.23.6
pytest-timeout==2.3.1
httpx==0.27.0           # for FastAPI test client

# Code quality
black==24.4.2
flake8==7.0.0
isort==5.13.2

# Visualisation (evaluation reports)
matplotlib==3.8.4

# Data validation
great-expectations==0.18.15

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
